Move mod/probe to src/Module/Probe
